+++ frontend/src/components/Item.tsx
@@ -3,32 +3,32 @@ import ListTitle from './ListTitle';
 import { SubscriptionUsedFront } from '../types';
 import { useNavigate } from 'react-router-dom';
 
-export type MainCardItemProps = {
+export type ItemProps = {
   title: string;
   SubscriptionList: SubscriptionUsedFront[];
 };
 
-export const StyledMainCardItem = styled.div`
+export const StyledItem = styled.div`
   display: flex;
   flex-direction: column;
   margin-bottom: 3rem;
 `;
 
-export const StyledMainCardItemSpan = styled.span`
+export const StyledItemSpan = styled.span`
   padding: 0.5rem 0;
   font-size: 0.9rem;
   cursor: pointer;
 `;
 
-const MainCardItem = ({ title, SubscriptionList }: MainCardItemProps) => {
+const Item = ({ title, SubscriptionList }: ItemProps) => {
   const navigate = useNavigate();
 
   return (
     <article>
       <ListTitle title={title} />
-      <StyledMainCardItem>
+      <StyledItem>
         {SubscriptionList.map((subscription, i) => (
-          <StyledMainCardItemSpan
+          <StyledItemSpan
             key={subscription.id}
             onClick={() =>
               navigate('/detail', {
@@ -37,11 +37,11 @@ const MainCardItem = ({ title, SubscriptionList }: MainCardItemProps) => {
             }
           >
             {i + 1}. {subscription.houseName}
-          </StyledMainCardItemSpan>
+          </StyledItemSpan>
         ))}
-      </StyledMainCardItem>
+      </StyledItem>
     </article>
   );
 };
 
-export default MainCardItem;
+export default Item;
